testing demo app without routing mixin
