+++ demo/src/components/dashboard-route.js
@@ -1,7 +1,6 @@
 import { LitElement, html, css } from 'lit';
-import {routingMixin} from '@jack-henry/web-component-router';
 
-class DashboardRoute extends routingMixin(LitElement) {
+class DashboardRoute extends LitElement {
   static styles = css`
     :host {
       display: block;
