+++ demo/src/components/section-b1-route.js
@@ -1,7 +1,6 @@
 import { LitElement, html, css } from 'lit';
-import {routingMixin} from '@jack-henry/web-component-router';
 
-class SectionB1Route extends routingMixin(LitElement) {
+class SectionB1Route extends LitElement {
   static styles = css`
     :host {
       display: block;
