+++ lib/animated-routing-mixin.js
@@ -7,7 +7,6 @@ import BasicRoutingInterface from './routing-interface.js';
  * @param {function(new:HTMLElement)} Superclass
  * @param {string} className
  * @mixinFunction
- * @polymer
  */
 function animatedRoutingMixin(Superclass, className) {
   /**
@@ -20,7 +19,6 @@ function animatedRoutingMixin(Superclass, className) {
 
   /**
    * @mixinClass
-   * @polymer
    * @extends {BasicRoutingElement}
    * @implements {BasicRoutingInterface}
    */
