+++ lib/route-tree-node.js
@@ -20,6 +20,7 @@
 
 import {Context} from './page.js';
 import RouteData from './route-data.js';
+import {loadRouteNode, removeRouteNode} from './route-change-handlers.js';
 import BasicRoutingInterface from './routing-interface.js';
 
 class RouteTreeNode {
@@ -291,10 +292,11 @@ class RouteTreeNode {
         const routingElem = /** @type {!BasicRoutingInterface} */ (
             /** @type {?} */ (currentExitNode.getValue().element)
         );
-        if (!routingElem.routeExit) {
-          throw new Error(`Element '${currentExitNode.getValue().tagName}' does not implement routeExit`);
+        if (routingElem.routeExit) {
+          await routingElem.routeExit(currentExitNode, nextExitNode, routeId, context);
+        } else {
+          await removeRouteNode(currentExitNode);
         }
-        await routingElem.routeExit(currentExitNode, nextExitNode, routeId, context);
         value.element = undefined;
       }
     }
@@ -308,10 +310,12 @@ class RouteTreeNode {
         const routingElem = /** @type {!BasicRoutingInterface} */ (
             /** @type {?} */ (currentEntryNode.getValue().element)
         );
-        if (!routingElem.routeEnter) {
-          throw new Error(`Element '${currentEntryNode.getValue().tagName}' does not implement routeEnter`);
+        let shouldContinue;
+        if(routingElem.routeEnter) {
+          shouldContinue = await routingElem.routeEnter(currentEntryNode, nextEntryNode, routeId, context);
+        } else {
+          shouldContinue = await loadRouteNode(currentEntryNode, nextEntryNode, routeId, context);
         }
-        const shouldContinue = await routingElem.routeEnter(currentEntryNode, nextEntryNode, routeId, context);
         if (shouldContinue === false) {
           break;
         }
