+++ package.json
@@ -37,7 +37,7 @@
     "@types/node": "^18.11.18"
   },
   "devDependencies": {
-    "@polymer/polymer": "^3.4.1",
+    "lit": "^3.1.1",
     "@types/node": "^22.10.0",
     "@vitest/browser": "^2.1.6",
     "playwright": "1.52.0",
